Former-commit-id: 2a9cfc8b8b9160d2e3def680fe8499259ca8859b

+++ db/migrate/20091118140336_add_duration_and_type_flag_to_tape.rb
@@ -1,42 +1,31 @@
 class AddDurationAndTypeFlagToTape < ActiveRecord::Migration
 
   def self.up
-
-    # remove old string-format duration (was nil, anyway)
-    remove_column :tapes, :duration
-
-    change_table :tapes do |t|
-      t.boolean :video
-      t.integer :duration
-    end
+  unless Project.name.to_sym == :mog
+    add_column :tapes, :video, :boolean
+    change_column :tapes, :duration, :integer
 
     # also change segment duration to int
-    remove_column :segments, :duration
-    add_column :segments, :duration, :integer
+    change_column :segments, :duration, :integer
 
     # add some more workflow-state flags to the interview
-    change_table :interviews do |t|
-      t.boolean :segmented, :default => false
-      t.boolean :researched, :default => false
-      t.boolean :proofread, :default => false
-    end
-
+    add_column :interviews, :segmented, :boolean, default: false
+    add_column :interviews, :researched, :boolean, default: false
+    add_column :interviews, :proofread, :boolean, default: false
 
     puts "setting to segmented on existing interviews"
-    Interview.all(
-        :joins => 'RIGHT JOIN tapes on tapes.interview_id = interviews.id',
-        :readonly => false
-    ).each do |interview|
-      interview.update_attribute :segmented, true
-      STDOUT::printf '.'
-      STDOUT.flush
-    end
+    #Interview.all(
+        #:joins => 'RIGHT JOIN tapes on tapes.interview_id = interviews.id',
+        #:readonly => false
+    #).each do |interview|
+    Interview.joins(:tapes).update_all(segmented: true)
     puts
 
   end
+  end
 
   def self.down
-
+  unless Project.name.to_sym == :mog
     remove_column :tapes, :video
     remove_column :tapes, :duration
     remove_column :segments, :duration
@@ -48,6 +37,7 @@ class AddDurationAndTypeFlagToTape < ActiveRecord::Migration
     remove_column :interviews, :researched
     remove_column :interviews, :proofread
 
+  end
   end
 
 end
