+++ db/migrate/20110308121855_add_interview_id_to_segments.rb
@@ -1,6 +1,7 @@
 class AddInterviewIdToSegments < ActiveRecord::Migration
 
   def self.up
+  unless Project.name.to_sym == :mog
 
     change_table :segments do |t|
       t.integer :interview_id
@@ -12,9 +13,12 @@ class AddInterviewIdToSegments < ActiveRecord::Migration
     end
 
   end
+  end
 
   def self.down
+  unless Project.name.to_sym == :mog
     remove_column :segments, :interview_id
+  end
   end
 
 end
