+++ db/migrate/20110310141001_create_location_segments.rb
@@ -1,6 +1,7 @@
 class CreateLocationSegments < ActiveRecord::Migration
 
   def self.up
+  unless Project.name.to_sym == :mog
 
     create_table :location_segments do |t|
       t.integer :location_reference_id
@@ -9,12 +10,15 @@ class CreateLocationSegments < ActiveRecord::Migration
     end
 
     add_index :location_segments, :location_reference_id
-    add_index :location_segments, [:location_reference_id, :interview_id]
+    add_index :location_segments, [:location_reference_id, :interview_id], name: 'location_segments_reference_and_interview'
 
+  end
   end
 
   def self.down
+  unless Project.name.to_sym == :mog
     drop_table :locations_segments
+  end
   end
 
 end
